+++ src/FloorOverlay.tsx
@@ -10,8 +10,8 @@ import { Component, createRef, h } from "preact";
 import Overlay from "./Overlay";
 import OverlaySearchBar from "./OverlaySearchBar";
 import { css, cx, mixins, theme } from "./style";
-import { FloorData } from "./data";
 import { createSearchMatcher, uiText } from "./util";
+import { FloorData } from "./web-sdk";
 
 export interface FloorOverlayProps {
   toggleFloorOverlay: (options: { open: boolean }) => void;
