+++ src/OverlayLayer.tsx
@@ -11,10 +11,10 @@ import OverlayPolyline from "./OverlayPolyline";
 import { css } from "./style";
 import { CustomOverlay } from "./web-sdk";
 
-type OverlayLayerProps = {
+interface OverlayLayerProps {
   mapZoomFactor: number;
   overlays: CustomOverlay[];
-};
+}
 
 const OverlayLayer: FunctionComponent<OverlayLayerProps> = ({
   mapZoomFactor,
